Merge skills and talents using IDs
For skills, attempt to find a target skill using IDs first (rather than
indices), then resort to comparing names. If a skill with matching name
is found, accept it as the target only if it doesn't have an ID yet.

For talents, check if a target talent exists at the same level and index
as the source talent.
    If so, compare their IDs (if the target talent has one) or names (if
    the target talent has no ID), but not both.
        If the ID/name match fails, discard the target talent and proceed
        as though there is no match.
    Otherwise, search through all talents, looking for:
        If the target talent has an ID, accept it if the ID matches.
        Do NOT compare their names.
        Otherwise, accept it if it has a matching name.

In general, this ensures that targets with IDs are always matched using
IDs, while those without IDs are matched by names only.