Match page URL when adding context menu in Chrome
For some reason, the WYSIWYG iframe's URL is now identical to the parent
page URL (Previously, it was always 'about:blank'). Because of this, the
context menu no longer appears in Chrome.

To fix, change documentUrlPatterns to always match the page URL. Note
that this was already the case in Firefox.