Use synchronous fs methods in merge-hdp.js
Since merge-hdp.js is a console script, readFileAsync() and
writeFileAsync() do not provide any discernable benefit over writeFile()
and readFile(). Also, the synchronous versions provide better stack
trace messages when an error occurs.

The "main" program code has been moved to the bottom of the script, so
that it can use constant variables defined before it.