fix(ts): Remove broken import aliasing trick
When I was converting the modules to ESM format, I came up with a weird
trick to "fix" the type acquisition of vendored modules, such as Preact
and htm. Clearly, it was broken.

Since I can now pull in the original type definitions for the vendored
modules using stub .d.ts files, the type checking works as intended,
and we have no need for these ugly hacks.