feat: Add createHdpHeroDataKoSchema()
This function takes the JSON schema for the English version of
herodata.json, and converts it to a schema for the Korean version.
Since the English and Korean schemas share many common elements, we can
generate a Korean schema on the fly instead of maintaining two similar
copies of the schema.

(herodata.json refers to JSON files generated by HeroesDataParser.)