refactor: Separate skill and talent schemas
Previously, the definition for "skill" in hots-schema.json reused the
"talent" definition with an extra clause: no "upgradeFor". While this
was good enough for validation, it could not be translated to a valid
Skill interface in TypeScript, because json-schema-to-typescript is not
smart enough for this.

To solve this, I made the "skill" definition entirely separate from
"talent". To avoid duplication, I used $refs for all members of "skill".