test: Fix Node.js environment detection
Since we are using jsdom-global to inject objects like 'window' and
'document' into the global object, we can't rely on the absence of a
'window' object to check if mocks.js is running inside Node.js.
Instead, use the presence of the 'global' object to check if we're
running in Node.js.

Note: JSDOM does not provide the Fetch API (`window.fetch()`), so we
still have to use the built-in `fs` module for testing in Node.js.