refactor: MultiSelectIcons.onSelectChange receives object
MultiSelectIcons.onSelectChange used to receive an array of selected
option IDs. This was unwieldly, as the status of each option had to be
manually checked with Array.prototype.includes().
Now, the callback simply receives an object that has each option ID as
keys, and the option's selection status (true/false) as values.