refactor: <MultiSelectIcons> is fully controlled
<MultiSelectIcons> is now a fully controlled component. It no longer
stores the state of checkboxes as internal state, and the checkboxes now
depend entirely on its props. Since it no longer has any state, it has
been converted to a function component.

Now, <HeroMenuFiltered> stores and controls the state of the checkboxes.
It had already been doing this before its code was extracted from
<DialogContent>; this commit merely ensures that we have a single source
of truth for the checkbox state.
Nevertheless, since I now have to pass the state of checkboxes as the
props to <MultiSelectIcons>, I also modified <HeroMenuFiltered>.

This also fixes a long-standing issue in <MultiSelectIcons>: Previously,
it would fire onSelectChange() only after its internal state had been
updated. This was causing unnecessary delays for the change to propagate
(and thus requiring extra awaits in tests), and also probably caused
unnecessary re-renders. Since we have completely eliminated internal
state in <MultiSelectIcons>, we can also remove this hack in the test.

I also updated the test file for <MultiSelectIcons> to properly check
the highlight status of the icons, and updated the expected values
passed to onSelectChange().