feat: <MultiSelectIcons> uses keyboard-accessible buttons
<MultiSelectIcons> uses <button> instead of <img>, which makes the icons
keyboard-accessible--can be focused using Tab, and clicked using Space
or Enter. Each button now acts as a toggle button (has the aria-pressed
attribute).

To avoid creating outlines on the Microtip tooltips, add a custom yellow
border (box-shadow) when the buttons are focused. Also, display the
tooltips when a button is hovered or active, but not when focused.